mais alguns fixes...

+++ src/game/base/gametile.cc
@@ -22,7 +22,7 @@ namespace base {
 
 const Vector2D GameTile::TILE_SIZE = Vector2D(23.0, 23.0);
 
-GameTile::GameTile(size_t x, size_t y) : node_(new Node) {
+GameTile::GameTile(int x, int y) : node_(new Node) {
 	ground_ = new TexturedRectangle( ResourceManager::GetTextureFromFile("data/white.png") );
 	node_->set_drawable(ground_);
     node_->modifier()->set_scale(Vector2D(0.382));
